requirements.txt
streamlit
openai
python-dotenv
textblob
speechrecognition
pyaudio
gtts
pyttsx3

#wheel @ file:///usr/local/Cellar/python%403.12/3.12.4/libexec/wheel-0.43.0-py3-none-any.whl#sha256=ff23205a590f2b902a2a21151019e96b93ab4a1409d7c9748b6bd210f4c0f44d
